Wrote code to save shader compile errors and warnings to a file when in debug mode. Added a project3 function to get the projected depth as well as screen position of a point in space. Changed the cockpit screen overlay. Fixed several warnings in shader files. Added an improved explosion flash effect. Improved plane smoke trails (still in progress) Added code to render the sun. Improved particle for explosion fire. Started writing an ephemeris class to compute the location of the sun. (and possibly moon, stars, etc.)

+++ Angle.h
@@ -105,10 +105,10 @@ public:
 	Angle operator+(Angle a){return Angle(ang+a.ang);}
 	Angle operator-(Angle a){return Angle(ang-a.ang);}
 
-	Angle operator*(double a){return Angle(ang*(float)a);}
-	Angle operator/(double a){return Angle(ang/(float)a);}
-	Angle operator+(double a){return Angle(ang+(float)a);}
-	Angle operator-(double a){return Angle(ang-(float)a);}
+	Angle operator*(double a){return Angle(ang*a);}
+	Angle operator/(double a){return Angle(ang/a);}
+	Angle operator+(double a){return Angle(ang+a);}
+	Angle operator-(double a){return Angle(ang-a);}
 
 	Angle operator*(float a){return Angle(ang*a);}
 	Angle operator/(float a){return Angle(ang/a);}
@@ -125,10 +125,10 @@ public:
 	Angle& operator+=(const Angle& a){ang+=a.ang; normalize(); return *this;}
 	Angle& operator-=(const Angle& a){ang-=a.ang; normalize(); return *this;}
 
-	Angle& operator*=(const double& a){ang*=(float)a; normalize(); return *this;}
-	Angle& operator/=(const double& a){ang/=(float)a; normalize(); return *this;}
-	Angle& operator+=(const double& a){ang+=(float)a; normalize(); return *this;}
-	Angle& operator-=(const double& a){ang-=(float)a; normalize(); return *this;}
+	Angle& operator*=(const double& a){ang*=a; normalize(); return *this;}
+	Angle& operator/=(const double& a){ang/=a; normalize(); return *this;}
+	Angle& operator+=(const double& a){ang+=a; normalize(); return *this;}
+	Angle& operator-=(const double& a){ang-=a; normalize(); return *this;}
 
 	Angle& operator*=(const float& a){ang*=a; normalize(); return *this;}
 	Angle& operator/=(const float& a){ang/=a; normalize(); return *this;}
